moved Diable2FA.tsx
moved Enable2FA.tsx

moved TwoFaContent.tsx

moved them to 2fa folder
