+++ frontend/src/components/settings/2fa/Enable2FA.tsx
@@ -1,12 +1,12 @@
 import useEnable2FA from "@/hooks/authentication/useEnable2FA";
-import { Button } from "../ui/button";
+import { Button } from "../../ui/button";
 
-const Enable = () => {
+const Enable2FA = () => {
 	const twoFaMutation = useEnable2FA();
 
 	return (
 		<div className="flex flex-col justify-center items-center space-y-12">
-			<p className="text-center">
+			<p className="text-center text-xs">
 				Enhance the security of your account by enabling Two-Factor Authentication (2FA). With 2FA,
 				an additional layer of protection is added to your login process, making it significantly
 				more secure click the button &quot;Enable 2fa&quot; and follow the steps
@@ -16,10 +16,10 @@ const Enable = () => {
 				disabled={twoFaMutation.isPending}
 				className="w-1/2"
 			>
-				Enable 2fa
+				Enable 2FA
 			</Button>
 		</div>
 	);
 };
 
-export default Enable;
+export default Enable2FA;
